feat(react): add compiled mode for preact upstream tests and extract shared transform plugin
Add dual-mode testing (no-compiler + with-compiler) for preact upstream tests
to verify that the SWC snapshot transform produces identical DOM output.

Extract transformReactLynxPlugin() from testing-library/src/vitest.config.js
into a standalone shared module (transformReactLynxPlugin.js) so both
preact-upstream-tests and testing-library can reuse the same plugin without
duplication.
